Refactored to a TGA header constructor
