+++ src/tga/encoder.rs
@@ -1,12 +1,10 @@
 use byteorder::{LittleEndian, WriteBytesExt};
-use std::{convert::TryFrom, io::Write};
+use std::io::Write;
 
-use super::image_type::{ImageType, ALPHA_BIT_MASK, SCREEN_ORIGIN_BIT_MASK};
+use super::header::Header;
 use crate::color::ColorType;
-use crate::error::{
-    ImageResult, ParameterError, ParameterErrorKind, UnsupportedError, UnsupportedErrorKind,
-};
-use crate::{image::ImageEncoder, ImageError, ImageFormat};
+use crate::error::ImageResult;
+use crate::image::ImageEncoder;
 
 /// TGA encoder.
 pub struct TgaEncoder<W: Write> {
@@ -22,7 +20,7 @@ impl<W: Write> TgaEncoder<W> {
     /// Encodes the image ```buf``` that has dimensions ```width```
     /// and ```height``` and ```ColorType``` ```color_type```.
     ///
-    /// The dimensions of the image must be between 1 and 65535 (inclusive) or
+    /// The dimensions of the image must be between 0 and 65535 (inclusive) or
     /// an error will be returned.
     pub fn encode(
         mut self,
@@ -31,23 +29,21 @@ impl<W: Write> TgaEncoder<W> {
         height: u32,
         color_type: ColorType,
     ) -> ImageResult<()> {
-        let (num_alpha_bits, other_channel_bits, image_type) = get_pixel_info(color_type)?;
-        let (width, height) = tga_dimensions(width, height)?;
+        let header = Header::from_pixel_info(color_type, width, height)?;
 
-        // Write TGA header.
-        self.writer.write_u8(0)?; // No ID string.
-        self.writer.write_u8(0)?; // No color map.
-        self.writer.write_u8(image_type as u8)?; // Image type.
-        self.writer.write_u16::<LittleEndian>(0)?; // No color map.
-        self.writer.write_u16::<LittleEndian>(0)?; // No color map.
-        self.writer.write_u8(0)?; // No color map.
-        self.writer.write_u16::<LittleEndian>(0)?; // X-origin.
-        self.writer.write_u16::<LittleEndian>(0)?; // Y-origin.
-        self.writer.write_u16::<LittleEndian>(width)?; // Width.
-        self.writer.write_u16::<LittleEndian>(height)?; // Height.
-        self.writer.write_u8(num_alpha_bits + other_channel_bits)?; // Bits per pixel.
-        self.writer
-            .write_u8((num_alpha_bits & ALPHA_BIT_MASK) | SCREEN_ORIGIN_BIT_MASK)?; // Upper left origin.
+        // Write out TGA header.
+        self.writer.write_u8(header.id_length)?;
+        self.writer.write_u8(header.map_type)?;
+        self.writer.write_u8(header.image_type)?;
+        self.writer.write_u16::<LittleEndian>(header.map_origin)?;
+        self.writer.write_u16::<LittleEndian>(header.map_length)?;
+        self.writer.write_u8(header.map_entry_size)?;
+        self.writer.write_u16::<LittleEndian>(header.x_origin)?;
+        self.writer.write_u16::<LittleEndian>(header.y_origin)?;
+        self.writer.write_u16::<LittleEndian>(header.image_width)?;
+        self.writer.write_u16::<LittleEndian>(header.image_height)?;
+        self.writer.write_u8(header.pixel_depth)?;
+        self.writer.write_u8(header.image_desc)?;
 
         // Write out Bgr(a)8 or L(a)8 image data.
         let mut image = Vec::from(buf);
@@ -76,37 +72,6 @@ impl<W: Write> ImageEncoder for TgaEncoder<W> {
     ) -> ImageResult<()> {
         self.encode(buf, width, height, color_type)
     }
-}
-
-fn get_pixel_info(color_type: ColorType) -> ImageResult<(u8, u8, ImageType)> {
-    let sizes = match color_type {
-        ColorType::Rgba8 | ColorType::Bgra8 => (8, 24, ImageType::RawTrueColor),
-        ColorType::Rgb8 | ColorType::Bgr8 => (0, 24, ImageType::RawTrueColor),
-        ColorType::La8 => (8, 8, ImageType::RawGrayScale),
-        ColorType::L8 => (0, 8, ImageType::RawGrayScale),
-        _ => {
-            return Err(ImageError::Unsupported(
-                UnsupportedError::from_format_and_kind(
-                    ImageFormat::Tga.into(),
-                    UnsupportedErrorKind::Color(color_type.into()),
-                ),
-            ))
-        }
-    };
-
-    Ok(sizes)
-}
-
-fn tga_dimensions(width: u32, height: u32) -> ImageResult<(u16, u16)> {
-    fn inner_dimensions(width: u32, height: u32) -> Option<(u16, u16)> {
-        let width = u16::try_from(width).ok()?;
-        let height = u16::try_from(height).ok()?;
-        Some((width, height))
-    }
-
-    inner_dimensions(width, height).ok_or(ImageError::Parameter(ParameterError::from_kind(
-        ParameterErrorKind::DimensionMismatch,
-    )))
 }
 
 #[cfg(test)]
